Renamed main.pyw to __main__.py, it can now be launched with:
pythonw -m runviewer.

Added a shortcut pointing to this

Made into a package with an __init__.py, added the capability for multiple
platforms to be supported. Setup.py creates an extension file named specifically
for the current platform, and __init__.py renames whichever extension file is
the right one for its bit-count to simply resample.pyd (win) or
resample.so (linux). Win 32 and 64 are included at the moment.
__init__.py will also call setup.py if the extenstions aren't compiled.